backend/foodgram/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-base64==2.0
flake8==5.0.4
idna==3.3
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.0
Pillow==9.2.0
psycopg2-binary==2.8.6
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.4.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.2.1
reportlab==3.6.11
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.11
